+++ x2engine/protected/controllers/SiteController.php
@@ -1,42 +1,38 @@
 <?php
-/*********************************************************************************
- * The X2CRM by X2Engine Inc. is free software. It is released under the terms of 
- * the following BSD License.
- * http://www.opensource.org/licenses/BSD-3-Clause
+/*****************************************************************************************
+ * X2CRM Open Source Edition is a customer relationship management program developed by
+ * X2Engine, Inc. Copyright (C) 2011-2013 X2Engine Inc.
  * 
- * X2Engine Inc.
- * P.O. Box 66752
- * Scotts Valley, California 95067 USA
+ * This program is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Affero General Public License version 3 as published by the
+ * Free Software Foundation with the addition of the following permission added
+ * to Section 15 as permitted in Section 7(a): FOR ANY PART OF THE COVERED WORK
+ * IN WHICH THE COPYRIGHT IS OWNED BY X2ENGINE, X2ENGINE DISCLAIMS THE WARRANTY
+ * OF NON INFRINGEMENT OF THIRD PARTY RIGHTS.
  * 
- * Company website: http://www.x2engine.com 
- * Community and support website: http://www.x2community.com 
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public License for more
+ * details.
  * 
- * Copyright (C) 2011-2012 by X2Engine Inc. www.X2Engine.com
- * All rights reserved.
+ * You should have received a copy of the GNU Affero General Public License along with
+ * this program; if not, see http://www.gnu.org/licenses or write to the Free
+ * Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+ * 02110-1301 USA.
  * 
- * Redistribution and use in source and binary forms, with or without modification, 
- * are permitted provided that the following conditions are met:
+ * You can contact X2Engine, Inc. P.O. Box 66752, Scotts Valley,
+ * California 95067, USA. or at email address contact@x2engine.com.
  * 
- * - Redistributions of source code must retain the above copyright notice, this 
- *   list of conditions and the following disclaimer.
- * - Redistributions in binary form must reproduce the above copyright notice, this 
- *   list of conditions and the following disclaimer in the documentation and/or 
- *   other materials provided with the distribution.
- * - Neither the name of X2Engine or X2CRM nor the names of its contributors may be 
- *   used to endorse or promote products derived from this software without 
- *   specific prior written permission.
+ * The interactive user interfaces in modified source and object code versions
+ * of this program must display Appropriate Legal Notices, as required under
+ * Section 5 of the GNU Affero General Public License version 3.
  * 
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
- * IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
- * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, 
- * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, 
- * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF 
- * LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE 
- * OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED 
- * OF THE POSSIBILITY OF SUCH DAMAGE.
- ********************************************************************************/
+ * In accordance with Section 7(b) of the GNU Affero General Public License version 3,
+ * these Appropriate Legal Notices must retain the display of the "Powered by
+ * X2Engine" logo. If the display of the logo is not reasonably feasible for
+ * technical reasons, the Appropriate Legal Notices must display the words
+ * "Powered by X2Engine".
+ *****************************************************************************************/
 
 /**
  * Primary/default controller for the web application.
@@ -64,20 +60,20 @@ class SiteController extends x2base {
 	public function accessRules() {
 		return array(
 			array('allow',
-				'actions'=>array('login','index','logout','warning','captcha','googleLogin'),
+				'actions'=>array('login','index','logout','warning','captcha','googleLogin','error'),
 				'users'=>array('*'),
 			),
 			array('allow', // allow authenticated user to perform 'create' and 'update' actions
 				'actions'=>array('groupChat','newMessage','getMessages','checkNotifications','updateNotes','addPersonalNote',
 				'getNotes','getURLs','addSite','deleteMessage','fullscreen','pageOpacity','widgetState','widgetOrder','saveGridviewSettings','saveFormSettings',
-					'saveWidgetHeight','inlineEmail','tmpUpload','upload','uploadProfilePicture','index','error','contact',
+					'saveWidgetHeight','inlineEmail','tmpUpload','upload','uploadProfilePicture','index','contact',
                     'viewNotifications','inlineEmail', 'toggleShowTags', 'appendTag', 'removeTag', 'addRelationship', 'createRecords',
                     'whatsNew','toggleVisibility','page', 'showWidget', 'hideWidget', 'reorderWidgets', 'minimizeWidget','publishPost','getEvents','loadComments',
-                    'loadPosts','addComment','flagPost','sendErrorReport','minimizePosts','bugReport','deleteRelationship'),
+                    'loadPosts','addComment','flagPost','sendErrorReport','minimizePosts','bugReport','deleteRelationship','toggleFeedControls','toggleFeedFilters'),
 				'users'=>array('@'),
 			),
 			 array('allow',
-				 'actions'=>array('motd'),
+				 'actions'=>array('motd','stickyPost'),
 				 'users'=>array('admin'),
 			 ),
 			array('deny', 
@@ -175,6 +171,9 @@ class SiteController extends x2base {
                 }
             }
             unset($_SESSION['feed-condition']);
+            if(!isset($_GET['filters'])){
+                unset($_SESSION['filters']);
+            }
             if(isset(Yii::app()->params->profile->defaultFeedFilters)){
                 $_SESSION['filters']=json_decode(Yii::app()->params->profile->defaultFeedFilters,true);
             }
@@ -291,15 +290,14 @@ class SiteController extends x2base {
             if(isset($_SESSION['lastEventId'])){
                 $condition.=" AND id <= ".$_SESSION['lastEventId'];
             }
-            $total=Events::model()->count();
-            $pages = new CPagination($total);
-            $pages->pageSize = 20;
             $dataProvider=new CActiveDataProvider('Events',array(
                 'criteria'=>array(
                     'condition'=>$condition,
                     'order'=>'timestamp DESC, id DESC',
                 ),
-                'pagination'=>$pages,
+                'pagination'=>array(
+                    'pageSize'=>20
+                ),
             ));
             $data=$dataProvider->getData();
             if(isset($data[count($data)-1]))
@@ -307,15 +305,26 @@ class SiteController extends x2base {
             else
                 $firstId=0;
             $users=User::getUserIds();
+            $_SESSION['stickyFlag']=false;
+            $_SESSION['firstFlag']=true;
+            $stickyDataProvider=new CActiveDataProvider('Events',array(
+                'criteria'=>array(
+                    'condition'=>'sticky=1',
+                    'order'=>'timestamp DESC, id DESC',
+                ),
+                'pagination'=>array(
+                    'pageSize'=>20
+                ),
+            ));
 			$this->render('whatsNew',array(
 				'dataProvider'=>$dataProvider,
                 'users'=>$users,
                 'lastEventId'=>!empty($lastId)?$lastId:0,
                 'firstEventId'=>!empty($firstId)?$firstId:0,
-                'lastTimestamp'=>$lastTimestamp
+                'lastTimestamp'=>$lastTimestamp,
+                'stickyDataProvider'=>$stickyDataProvider,
 			));
-		}
-		else{
+		}else{
 			$this->redirect('login');
 		}
 	}
@@ -507,6 +516,38 @@ class SiteController extends x2base {
         }
     }
     
+    public function actionStickyPost($id){
+        $event=X2Model::model('Events')->findByPk($id);
+        if(isset($event)){
+            $event->sticky=!$event->sticky;
+            $event->update();
+        }
+    }
+    
+    public function actionToggleFeedControls(){
+        $profile=Yii::app()->params->profile;
+        if(isset($profile)){
+            $profile->fullFeedControls=!$profile->fullFeedControls;
+            $profile->update();
+        }
+    }
+    
+    public function actionToggleFeedFilters($filter){
+        $profile=Yii::app()->params->profile;
+        if(isset($profile)){
+            $filters=json_decode($profile->feedFilters,true);
+            if(isset($filters[$filter])){
+                $filters[$filter]=$filters[$filter]==1?0:1;
+            }else{
+                $filters[$filter]=0;
+            }
+            $flag=$filters[$filter];
+            $profile->feedFilters=json_encode($filters);
+            $profile->update(array('feedFilters'));
+            echo $flag;
+        }
+    }
+    
     public function actionMinimizePosts(){
         if(isset($_GET['minimize'])){
             $profile=Yii::app()->params->profile;
@@ -659,7 +700,8 @@ class SiteController extends x2base {
 	 */
 	public function actionDeleteMessage($id,$url){
 		$note=Social::model()->findByPk($id);
-		$note->delete();
+        if(isset($note))
+            $note->delete();
 		$this->redirect($url);
 	} 
 
@@ -1015,8 +1057,7 @@ class SiteController extends x2base {
 			$model->fileName=$name;
 			
 			// download and save picture
-			$img = file_get_contents($photourl);
-			file_put_contents('uploads/'.$name, $img);
+			$img = FileUtil::ccopy($photourl,"uploads/$name");
 			$model->save();
 			
 			// put picture into new action
@@ -1095,13 +1136,13 @@ class SiteController extends x2base {
 					if($module->exists)
 						$this->redirect(array($profile->startPage.'/'.$profile->startPage.'/index'));
 				} else {
-					$page=DocChild::model()->findByAttributes(array('title'=>ucfirst($profile->startPage)));
+					$page = CActiveRecord::model('Docs')->findByAttributes(array('name'=>ucfirst($profile->startPage)));
 					if(isset($page)) {
 						$id=$page->id;
-						$menuItems[$key] = array('label' =>ucfirst($value),'url' => array('/admin/viewPage/'.$id),'active'=>Yii::app()->request->requestUri==Yii::app()->request->scriptUrl.'/admin/viewPage/'.$id?true:null);
+						$this->redirect('docs/'.$id.'?static=true');
 				
 					} else {
-					$this->redirect(array('site/whatsNew'));
+                        $this->redirect(array('site/whatsNew'));
 					}
 				}
 			}
@@ -1212,6 +1253,8 @@ class SiteController extends x2base {
                     'phpversion'=>$phpversion,
                     'x2version'=>$x2version,
                     'email'=>$email,
+                    'user'=>Yii::app()->user->getName(),
+                    'isAdmin'=>Yii::app()->user->checkAccess('AdminIndex'),
                 ))));
                 
                 $errorReport=base64_encode(serialize(array_merge($error,array(
@@ -1222,6 +1265,8 @@ class SiteController extends x2base {
                     'phpversion'=>$phpversion,
                     'x2version'=>$x2version,
                     'email'=>$email,
+                    'user'=>Yii::app()->user->getName(),
+                    'isAdmin'=>Yii::app()->user->checkAccess('AdminIndex'),
                 ))));
                 
 				$this->render('error', array_merge($error,array(
@@ -1254,12 +1299,16 @@ class SiteController extends x2base {
             'phpversion'=>$phpversion,
             'x2version'=>$x2version,
             'adminEmail'=>$email,
+            'user'=>Yii::app()->user->getName(),
+            'isAdmin'=>Yii::app()->user->checkAccess('AdminIndex'),
         )));
 
         $errorReport=base64_encode(serialize(array(
             'phpversion'=>$phpversion,
             'x2version'=>$x2version,
             'adminEmail'=>$email,
+            'user'=>Yii::app()->user->getName(),
+            'isAdmin'=>Yii::app()->user->checkAccess('AdminIndex'),
         )));
         
         $this->render('bugReport',array(
@@ -1349,7 +1398,6 @@ class SiteController extends x2base {
 	 */
 	public function actionLogin() {
 		$this->layout = '//layouts/login';
-		
 		if(Yii::app()->user->isInitialized && !Yii::app()->user->isGuest) {
 			$this->redirect(Yii::app()->homeUrl);
 			return;
@@ -1521,7 +1569,7 @@ class SiteController extends x2base {
 					
 					Yii::app()->session['loginTime']=time();
 						$session->status=1;
-
+					
 					if(Yii::app()->user->returnUrl=='site/index')
 						$this->redirect('index');
 					else
@@ -1555,91 +1603,37 @@ class SiteController extends x2base {
 	}
 	
 	/**
-	 * Add a tag to a module.
+	 * Adds a tag to a model.
 	 * 
-	 * Echoes true if tag was created (and was not a duplicate)
+	 * Echoes "true" if tag was created (and was not a duplicate), "false" otherwise.
 	 */
 	public function actionAppendTag() {
-		if( isset($_POST['Type']) && isset($_POST['Id']) && isset($_POST['Tag']) ) {
-			$type = ucfirst($_POST['Type']);
-			$id = $_POST['Id'];
-			$value = $_POST['Tag'];
+		if(isset($_POST['Type'],$_POST['Id'],$_POST['Tag']) && ctype_alpha($_POST['Type'])) {
+			$model = X2Model::model($_POST['Type'])->findByPk($_POST['Id']);
 			
-			if($type == 'Quotes' || $type == 'Products') // fix for products and quotes
-				$model = X2Model::model(rtrim($type, 's'))->findByPk($id);
-			elseif($type == 'Opportunities'){
-				$model = X2Model::model('Opportunity')->findByPk($id);
-                $type="Opportunity";
-            }else
-				$model = X2Model::model($type)->findByPk($id);
-			if($model) {
-				
-				// check for duplicate tag
-				$tag = Tags::model()->findByAttributes(array('type'=>$type, 'itemId'=>$id, 'tag'=>$value));
-				if($tag) {
-					echo json_encode(false); // tag was a duplicate
-					return;
-				} else {
-					// create tag
-					$tag = new Tags;
-					$tag->taggedBy = Yii::app()->user->name;
-					$tag->timestamp = time();
-					$tag->type = $type;
-					$tag->itemId = $id;
-					$tag->tag = $value;
-					
-					if($type == 'Contacts')
-						$tag->itemName = $model->firstName." ".$model->lastName;
-					else if($type == 'Actions')
-						$tag->itemName = $model->actionDescription;
-					else if($type == 'Docs')
-						$tag->itemName = $model->title;
-					else
-						$tag->itemName = $model->name;
-						
-					if($tag->save()) {
-						echo json_encode(true);
-						return;
-					}
-				}
+			if($model !== null && $model->addTags($_POST['Tag'])) {
+				echo 'true';
+				return;
 			}
 		}
-		json_encode(false);
+		echo 'false';
 	}
 	
 	/**
-	 * Remove a tag from a module.
+	 * Removes a tag from a model.
 	 * 
-	 * Echoes true if tag was removed.
+	 * Echoes "true" if tag was removed, "false" otherwise.
 	 */
 	public function actionRemoveTag() {
-		if( isset($_POST['Type']) && isset($_POST['Id']) && isset($_POST['Tag']) ) {
-			$type = ucfirst($_POST['Type']);
-			$id = $_POST['Id'];
-			$value = $_POST['Tag'];
-			if($type == 'Quotes' || $type == 'Products') // fix for products and quotes
-				$model = X2Model::model(rtrim($type, 's'))->findByPk($id);
-			elseif($type == 'Opportunities')
-				$model = X2Model::model('Opportunity')->findByPk($id);
-			else
-				$model = X2Model::model($type)->findByPk($id);
-			if($model) {
-				
-				// make sure tag exists
-				$tag = Tags::model()->findByAttributes(array('type'=>$type, 'itemId'=>$id, 'tag'=>$value));
-				if($tag) {
-					if($tag->delete()) {
-						echo json_encode(true); // tag was removed
-						return;
-					}
-				} else {
-					// tag doesn't exist
-					echo json_encode(false);
-					return;
-				}
+		if(isset($_POST['Type'],$_POST['Id'],$_POST['Tag']) && ctype_alpha($_POST['Type'])) {
+			$model = X2Model::model($_POST['Type'])->findByPk($_POST['Id']);
+			
+			if($model !== null && $model->removeTags($_POST['Tag'])) {
+				echo 'true';
+				return;
 			}
 		}
-		echo json_encode(false);
+		echo 'false';
 	}
 	
 	
@@ -1852,7 +1846,7 @@ class SiteController extends x2base {
 		$newVersion = '';
 		
 		foreach($updateSources as $url) {
-			$sourceVersion = @file_get_contents($url,0,$context);
+			$sourceVersion = FileUtil::getContents($url,0,$context);
 			if($sourceVersion !== false) {
 				$newVersion = $sourceVersion;
 				break;
@@ -1907,23 +1901,19 @@ class SiteController extends x2base {
 				X2Model::model('Session')->deleteByPk($_SESSION['sessionId']);
 			else
 				X2Model::model('Session')->deleteAllByAttributes(array('IP'=>$this->getRealIp()));
-				
-			// $session = Session::model()->findByAttributes(array('user'=>$user->username));
-			// $session = Session::model()->deleteAllByAttributes(array('user'=>$user->username));
-			// if(isset($session))
-				// $session->delete();
-			// $user->save();
+			
 		}
 		if(isset($_SESSION['access_token']))
 			unset($_SESSION['access_token']);
 		
 		Yii::app()->user->logout();
+		
 		$this->redirect(Yii::app()->homeUrl);
 	}
 
 	public function actionToggleVisibility(){
-		$ip = $this->getRealIp();
-		$session = Session::model()->findByAttributes(array('user'=>Yii::app()->user->getName(),'IP'=>$ip));
+        $id=$_SESSION['sessionId'];
+		$session = Session::model()->findByAttributes(array('id'=>$id));
 		if(isset($session)) {
 			$session->status = !$session->status;
 			$session->save();
